FSTech_Consulting_Agency/requirements.txt
openai
requests
python-dotenv
clickup-python
